refactor(gateway): feedback is a store, move it under storage, refactor(gateway): split core/runtime into turn, process and composition
