+++ .importlinter.strict
@@ -40,9 +40,9 @@ ignore_imports =
     platform.scheduler.executor -> integrations.telegram.delivery
     platform.scheduler.executor -> integrations.telegram.formatting
     # surfaces -> gateway (cli + repl gateway commands)
-    surfaces.cli.commands.gateway -> gateway.core.runtime.daemon
+    surfaces.cli.commands.gateway -> gateway.core.process.daemon
     surfaces.gateway_entry -> gateway.core.runtime.controller
-    surfaces.interactive_shell.command_registry.gateway_cmds -> gateway.core.runtime.daemon
+    surfaces.interactive_shell.command_registry.gateway_cmds -> gateway.core.process.daemon
     surfaces.interactive_shell.controller -> gateway.web.web_server
     # tools -> integrations (T-4 debt — vendor tools / reporting)
     tools.cross_vendor.fix_sentry_issue.context -> integrations.sentry
