+++ gateway/core/process/__init__.py
@@ -1,0 +1,1 @@
+"""The gateway process itself: supervision, readiness and background threads."""
