+++ gateway/core/process/daemon.py
