+++ gateway/core/process/polling_thread.py
