+++ gateway/core/storage/feedback/__init__.py
@@ -1,5 +1,5 @@
 """Transport-neutral feedback persistence."""
 
-from gateway.core.feedback.jsonl import append_feedback_entry
+from gateway.core.storage.feedback.jsonl import append_feedback_entry
 
 __all__ = ["append_feedback_entry"]
