+++ gateway/core/turn/turn_handler.py
@@ -26,12 +26,12 @@ from core.agent_harness.spi.session_goal import (
     format_session_goal_progress,
     format_session_goal_status_line,
 )
-from gateway.core.runtime.cancel_console import CancelConsole, ensure_turn_cancel
-from gateway.core.runtime.capability_policy import ensure_gateway_capability_policy
-from gateway.core.runtime.concurrency import TurnConcurrencyGate
-from gateway.core.runtime.session_agents import SessionAgentPool
-from gateway.core.runtime.status_messages import EMPTY_RESPONSE_MESSAGE
 from gateway.core.transport_api import GatewaySink
+from gateway.core.turn.cancel_console import CancelConsole, ensure_turn_cancel
+from gateway.core.turn.capability_policy import ensure_gateway_capability_policy
+from gateway.core.turn.concurrency import TurnConcurrencyGate
+from gateway.core.turn.session_agents import SessionAgentPool
+from gateway.core.turn.status_messages import EMPTY_RESPONSE_MESSAGE
 from platform.analytics.cli import (
     capture_gateway_turn_completed,
     capture_gateway_turn_failed,
