+++ gateway/tests/runtime/test_cancel_console.py
@@ -7,7 +7,7 @@ import threading
 
 from rich.console import Console
 
-from gateway.core.runtime.cancel_console import CancelConsole
+from gateway.core.turn.cancel_console import CancelConsole
 
 
 def _console() -> Console:
