+++ gateway/transports/buzz/startup.py
@@ -10,7 +10,7 @@ from __future__ import annotations
 
 import logging
 
-from gateway.core.runtime.polling_thread import PollingBackground
+from gateway.core.process.polling_thread import PollingBackground
 from gateway.core.transport_api import GatewayAgentCallback
 from gateway.transports.buzz.background import start_buzz_gateway_background
 from gateway.transports.buzz.runtime import (
