+++ gateway/transports/telegram/settings.py
@@ -11,8 +11,8 @@ from pydantic import Field, ValidationError, field_validator
 from pydantic_settings import BaseSettings, NoDecode, SettingsConfigDict
 
 from config.strict_config import StrictConfigModel
-from gateway.core.runtime.concurrency import turn_limit_for_profile
 from gateway.core.runtime.errors import GatewayConfigurationError
+from gateway.core.turn.concurrency import turn_limit_for_profile
 from integrations.messaging_security import MessagingIdentityPolicy, MessagingPlatform
 from integrations.store import get_integration
