+++ gateway/web/webapp.py
@@ -36,7 +36,7 @@ ensure_project_platform_package()
 
 from bootstrap.process import WEB_PROFILE, configure_process  # noqa: E402
 from core.agent_harness import AgentSession  # noqa: E402
-from gateway.core.runtime.readiness import is_gateway_ready  # noqa: E402
+from gateway.core.process.readiness import is_gateway_ready  # noqa: E402
 from gateway.core.storage import open_database  # noqa: E402
 from gateway.core.storage.investigations.repository import investigation_repository  # noqa: E402
 from gateway.web.investigations import router as investigations_router  # noqa: E402
@@ -214,7 +214,7 @@ def investigate(req: InvestigateRequest, request: Request) -> InvestigateRespons
     if (auth_error := _gateway_auth_error(request)) is not None:
         return auth_error
 
-    from gateway.core.runtime.concurrency import process_turn_gate
+    from gateway.core.turn.concurrency import process_turn_gate
 
     gate = process_turn_gate()
     # Same process gate as chat / scheduler — busy-drop like GatewayTurnHandler.
