+++ surfaces/interactive_shell/command_registry/gateway_cmds.py
@@ -5,7 +5,7 @@ from __future__ import annotations
 from rich.console import Console
 from rich.markup import escape
 
-from gateway.core.runtime.daemon import (
+from gateway.core.process.daemon import (
     GATEWAY_LOG_FILE,
     gateway_daemon_pid,
     read_component_status,
