+++ tests/shared/test_surface_border.py
@@ -39,7 +39,7 @@ _CLI_IMPORTS_FROM_SHELL: frozenset[str] = frozenset()
 _SURFACES_IMPORT_FROM_GATEWAY: frozenset[str] = frozenset(
     {
         # `opensre gateway start|stop|status` and the REPL's /gateway commands.
-        "gateway.core.runtime.daemon",
+        "gateway.core.process.daemon",
         # The console entry that constructs the process and starts it.
         "gateway.core.runtime.controller",
         # The REPL hosts the web app in-thread for local use.
